refactor SharepointClient -> (SharepointClient, SharepointTableProvider)
